web/requirements.txt
Django==1.11.16
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.3.0
psycopg2==2.7.5
pytz==2018.5
requests==2.18.4
simplejson==3.16.0
urllib3==1.22
celery==4.2.1
redis==2.10.6
django-cms==3.6.0
django-filter==2.1.0
djangocms-text-ckeditor==3.6.0
django-webpack-loader==0.6.0
djangorestframework-filters==1.0.0.dev0
django-admin-sortable2==0.7.1
django-autoslug==1.9.4
django-parler==1.9.2
djangocms-multimenus==0.2.1
django-redis==4.10.0
django-filer==1.5.0